end pirate anchor

+++ src/main/java/com/github/alexthe666/alexsmobs/block/BlockEndPirateAnchor.java
@@ -36,21 +36,21 @@ import java.util.Locale;
 public class BlockEndPirateAnchor extends BaseEntityBlock implements AMSpecialRenderBlock {
 
     public static final BooleanProperty EASTORWEST = BooleanProperty.create("eastorwest");
-    public static final EnumProperty<BlockEndPirateAnchor.PieceType> PIECE = EnumProperty.create("piece", BlockEndPirateAnchor.PieceType.class);
+    public static final BooleanProperty CENTER = BooleanProperty.create("center");
     protected static final VoxelShape FULL_AABB_EW = Block.box(0.0D, 0.0D, 4D, 16.0D, 16.0D, 12.0D);
     protected static final VoxelShape FULL_AABB_NS = Block.box(4.0D, 0.0D, 0.0D, 12.0D, 16.0D, 16.0D);
-    protected static final VoxelShape CHAIN_AABB = Block.box(4.0D, 0.0D, 4.0D, 12.0D, 16.0D, 12.0D);
 
     protected BlockEndPirateAnchor() {
         super(Properties.of(Material.STONE, MaterialColor.COLOR_BLACK).friction(0.97F).strength(10.0F).lightLevel((i) -> 6).sound(SoundType.STONE).noOcclusion());
-        this.registerDefaultState(this.stateDefinition.any().setValue(EASTORWEST, Boolean.valueOf(false)).setValue(PIECE, PieceType.ANCHOR));
+        this.registerDefaultState(this.stateDefinition.any().setValue(EASTORWEST, Boolean.valueOf(false)).setValue(CENTER, true));
     }
 
     public static boolean isClearForPlacement(LevelReader reader, BlockPos center, boolean eastOrWest) {
         List<BlockPos> offsets = TileEntityEndPirateAnchor.getValidBBPositions(eastOrWest);
         for (BlockPos offset : offsets) {
             BlockPos check = center.offset(offset);
-            if (!reader.isEmptyBlock(check) || !reader.getBlockState(check).getMaterial().isReplaceable()) {
+            BlockState state = reader.getBlockState(check);
+            if ((!reader.isEmptyBlock(check) || !state.getMaterial().isReplaceable()) && !state.is(AMBlockRegistry.END_PIRATE_ANCHOR_CHAIN.get())) {
                 return false;
             }
         }
@@ -61,24 +61,30 @@ public class BlockEndPirateAnchor extends BaseEntityBlock implements AMSpecialRe
         List<BlockPos> offsets = TileEntityEndPirateAnchor.getValidBBPositions(state.getValue(EASTORWEST));
         for (BlockPos offset : offsets) {
             if (!offset.equals(BlockPos.ZERO)) {
-                level.setBlock(pos.offset(offset), state.setValue(PIECE, PieceType.ANCHOR_SIDE), 2);
+                level.setBlock(pos.offset(offset), state.setValue(CENTER, false), 2);
             }
         }
     }
 
-    public static void removeAnchor(Level level, BlockPos pos, BlockState state) {
-        List<BlockPos> offsets = TileEntityEndPirateAnchor.getValidBBPositions(state.getValue(EASTORWEST));
+    public static void removeAnchor(Level level, BlockPos pos, boolean eastOrWest) {
+        List<BlockPos> offsets = TileEntityEndPirateAnchor.getValidBBPositions(eastOrWest);
         for (BlockPos offset : offsets) {
             level.setBlock(pos.offset(offset), Blocks.AIR.defaultBlockState(), 67);
         }
     }
 
+    public void onRemove(BlockState state, Level level, BlockPos blockPos, BlockState newState, boolean force) {
+        if (state.getValue(CENTER)) {
+            removeAnchor(level, blockPos, state.getValue(EASTORWEST));
+        }
+        super.onRemove(state, level, blockPos, newState, force);
+    }
+
+
     public BlockState getStateForPlacement(BlockPlaceContext context) {
         LevelReader levelreader = context.getLevel();
         BlockPos blockpos = context.getClickedPos();
         BlockPos actualPos = context.getClickedPos().relative(context.getClickedFace().getOpposite());
-        BlockPos u = blockpos.above();
-        BlockPos d = blockpos.below();
         BlockState clickState = levelreader.getBlockState(actualPos);
         boolean axis = context.getHorizontalDirection().getAxis() == Direction.Axis.X;
         if (clickState.getBlock() instanceof BlockEndPirateAnchor) {
@@ -87,42 +93,15 @@ public class BlockEndPirateAnchor extends BaseEntityBlock implements AMSpecialRe
         return isClearForPlacement(levelreader, blockpos, axis) ? defaultBlockState().setValue(EASTORWEST, axis) : null;
     }
 
-    public boolean isLadder(BlockState state, net.minecraft.world.level.LevelReader world, BlockPos pos, net.minecraft.world.entity.LivingEntity entity) {
-        return state.getValue(PIECE) == PieceType.CHAIN;
-    }
-
-    public boolean isScaffolding(BlockState state, LevelReader world, BlockPos pos, LivingEntity entity) {
-        return state.getValue(PIECE) == PieceType.CHAIN;
-    }
-
-    public void entityInside(BlockState state, Level level, BlockPos pos, Entity entity) {
-        if (entity instanceof LivingEntity && state.getValue(PIECE) == PieceType.CHAIN) {
-            LivingEntity livingEntity = (LivingEntity)entity;
-            if (livingEntity.horizontalCollision && !livingEntity.isInWater()) {
-                livingEntity.fallDistance = 0.0F;
-                Vec3 motion = livingEntity.getDeltaMovement();
-                double d0 = Mth.clamp(motion.x, -0.15F, 0.15F);
-                double d1 = Mth.clamp(motion.z, -0.15F, 0.15F);
-                double d2 = 0.3D;
-                if (d2 < 0.0D && livingEntity.isSuppressingSlidingDownLadder()) {
-                    d2 = 0.0D;
-                }
-                motion = new Vec3(d0, d2, d1);
-                livingEntity.setDeltaMovement(motion);
-            }
-        }
-    }
-
-
     public void neighborChanged(BlockState state, Level level, BlockPos pos, Block block, BlockPos p_52780_, boolean p_52781_) {
-        if (state.getValue(PIECE) == PieceType.ANCHOR_SIDE) {
+        if (!state.getValue(CENTER)) {
             for (int i = -2; i <= 2; i++) {
                 for (int j = -3; j <= 3; j++) {
                     for (int k = -2; k <= 2; k++) {
                         BlockPos offsetPos = pos.offset(i, j, k);
                         if (level.getBlockEntity(offsetPos) instanceof TileEntityEndPirateAnchor anchor) {
-                            if (!anchor.hasAllAnchorBlocks()) {
-                                removeAnchor(level, offsetPos, level.getBlockState(offsetPos));
+                            BlockState offsetState = level.getBlockState(offsetPos);
+                            if (!anchor.hasAllAnchorBlocks() && offsetState.is(this)) {
                                 level.destroyBlock(offsetPos, true);
                             }
                         }
@@ -140,27 +119,22 @@ public class BlockEndPirateAnchor extends BaseEntityBlock implements AMSpecialRe
     }
 
     public boolean canSurviveAnchor(BlockState state, LevelReader world, BlockPos pos) {
-        if (state.getValue(PIECE) == PieceType.ANCHOR) {
+        if (state.getValue(CENTER)) {
             return true;
-        } else if (state.getValue(PIECE) == PieceType.ANCHOR_SIDE) {
+        } else {
             for (int i = -1; i <= 1; i++) {
                 for (int j = -3; j <= 0; j++) {
                     for (int k = -1; k <= 1; k++) {
                         BlockPos offsetPos = pos.offset(i, j, k);
                         BlockState anchorState = world.getBlockState(offsetPos);
-                        if (anchorState.getBlock() instanceof BlockEndPirateAnchor && anchorState.getValue(PIECE) == PieceType.ANCHOR && isPartOfAnchor(anchorState, world, offsetPos, pos, state.getValue(EASTORWEST))) {
+                        if (anchorState.getBlock() instanceof BlockEndPirateAnchor && anchorState.getValue(CENTER) && isPartOfAnchor(anchorState, world, offsetPos, pos, state.getValue(EASTORWEST))) {
                             return true;
                         }
                     }
                 }
             }
-        } else if (state.getValue(PIECE) == PieceType.CHAIN) {
-            BlockPos below = pos.below();
-            BlockState chainBelow = world.getBlockState(below);
-            BlockState chainAbove = world.getBlockState(below);
-            return chainBelow.getBlock() instanceof BlockEndPirateAnchor && (chainAbove.getBlock() instanceof BlockEndPirateAnchor || chainAbove.getBlock() instanceof BlockEndPirateAnchorWinch);
+            return false;
         }
-        return false;
     }
 
     public boolean isPartOfAnchor(BlockState anchor, LevelReader level, BlockPos center, BlockPos pos, boolean eastOrWest) {
@@ -172,46 +146,29 @@ public class BlockEndPirateAnchor extends BaseEntityBlock implements AMSpecialRe
     }
 
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> p_58032_) {
-        p_58032_.add(EASTORWEST, PIECE);
+        p_58032_.add(EASTORWEST, CENTER);
     }
 
     public VoxelShape getShape(BlockState state, BlockGetter getter, BlockPos pos, CollisionContext context) {
-        if (state.getValue(PIECE) == PieceType.CHAIN) {
-            return CHAIN_AABB;
-        }
         return state.getValue(EASTORWEST) ? FULL_AABB_NS : FULL_AABB_EW;
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
-        return state.getValue(PIECE) == PieceType.ANCHOR ? new TileEntityEndPirateAnchor(pos, state) : null;
+        return state.getValue(CENTER) ? new TileEntityEndPirateAnchor(pos, state) : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level p_152180_, BlockState state, BlockEntityType<T> p_152182_) {
-        return state.getValue(PIECE) == PieceType.ANCHOR ? createTickerHelper(p_152182_, AMTileEntityRegistry.END_PIRATE_ANCHOR.get(), TileEntityEndPirateAnchor::commonTick) : null;
+        return state.getValue(CENTER) ? createTickerHelper(p_152182_, AMTileEntityRegistry.END_PIRATE_ANCHOR.get(), TileEntityEndPirateAnchor::commonTick) : null;
     }
 
     public RenderShape getRenderShape(BlockState state) {
-        return state.getValue(PIECE) == PieceType.ANCHOR_SIDE ? RenderShape.INVISIBLE : RenderShape.ENTITYBLOCK_ANIMATED;
+        return state.getValue(CENTER) ? RenderShape.ENTITYBLOCK_ANIMATED : RenderShape.INVISIBLE;
     }
 
     public List<ItemStack> getDrops(BlockState state, LootContext.Builder builder) {
-        return state.getValue(PIECE) == PieceType.ANCHOR ? super.getDrops(state, builder) : Collections.emptyList();
-    }
-
-    public enum PieceType implements StringRepresentable {
-        ANCHOR,
-        ANCHOR_SIDE,
-        CHAIN;
-
-        public String toString() {
-            return this.getSerializedName();
-        }
-
-        public String getSerializedName() {
-            return this.name().toLowerCase(Locale.ROOT);
-        }
+        return state.getValue(CENTER) ? super.getDrops(state, builder) : Collections.emptyList();
     }
 }
